Added some changes suggested by reviewers.

+++ students/rl/course_exercise_1/particle.f90
@@ -10,7 +10,7 @@ module particle
   type :: particle3d
      type(point3d) :: p  ! position
      type(vector3d) :: v  ! velocity
-     real :: m  ! mass
+     real(dp) :: m  ! mass
   end type particle3d
 
 end module particle
